created separte module to make url, api calls in two routes: / and /repos
